requirements.txt
numpy
pymysql
pandas
scipy
TA-Lib
technical
openpyxl
cryptography
discord.py
python-dotenv